Separate @adeira/eslint-runner to the new NPM package
So it can be used independently.

Closes: https://github.com/adeira/universe/issues/372
