+++ src/eslint-config-adeira/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
+# 1.2.0
+
 - Eslint runner no longer crashes if git is not initialised or if there is no git remote. It will fall back to lint all files instead.
+- Eslint runner has been moved the the `@adeira/eslint-runner` and will be eventually removed from this package.
 - Added `react/jsx-no-script-url` as next_version_error
 - `eslint-comments/no-unlimited-disable` rule is turned on, ignoring all `__generated__/*.graphql.js` files by default, so files autogenerated by Relay compiler are not reported. Follow [Eslint Docs](https://eslint.org/docs/user-guide/configuring#disabling-rules-only-for-a-group-of-files) on how to disable additional group of files if needed.
 
@@ -22,7 +25,7 @@
 
 # 1.1.0
 
-- New CLI option `--no-warnings`
+- New Eslint Runner CLI option `--no-warnings`
 
 # 1.0.0
