+++ src/eslint-config-adeira/package.json
@@ -4,15 +4,12 @@
   "homepage": "https://github.com/adeira/universe/tree/master/src/eslint-config-adeira",
   "license": "MIT",
   "private": false,
-  "version": "1.1.0",
+  "version": "1.2.0",
   "main": "index",
   "sideEffects": false,
   "dependencies": {
-    "@adeira/monorepo-utils": "^0.3.0",
-    "@babel/register": "^7.8.3",
+    "@adeira/eslint-runner": "^0.1.0",
     "@babel/runtime": "^7.8.4",
-    "chalk": "^3.0.0",
-    "create-jest-runner": "^0.5.3",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-adeira": "^0.7.0",
     "eslint-plugin-babel": "^5.3.0",
@@ -28,10 +25,7 @@
     "eslint-plugin-react": "^7.18.3",
     "eslint-plugin-react-hooks": "^2.3.0",
     "eslint-plugin-relay": "^1.5.0",
-    "is-ci": "^2.0.0",
-    "prettier": "^1.19.1",
-    "strip-ansi": "^6.0.0",
-    "text-table": "^0.2.0"
+    "prettier": "^1.19.1"
   },
   "devDependencies": {
     "deep-diff": "^1.0.2",
