+++ src/eslint-runner/.npmignore
@@ -1,0 +1,1 @@
+__tests__
