+++ src/eslint-runner/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@adeira/eslint-runner",
+  "description": "Eslint runner which is faster than Eslint itself 🤯.",
+  "homepage": "https://github.com/adeira/universe/tree/master/src/eslint-runner",
+  "license": "MIT",
+  "private": false,
+  "version": "0.1.0",
+  "main": "index",
+  "sideEffects": false,
+  "dependencies": {
+    "@adeira/monorepo-utils": "^0.2.0",
+    "@babel/register": "^7.8.3",
+    "@babel/runtime": "^7.8.4",
+    "chalk": "^3.0.0",
+    "create-jest-runner": "^0.5.3",
+    "strip-ansi": "^6.0.0",
+    "text-table": "^0.2.0",
+    "is-ci": "^2.0.0"
+  },
+  "peerDependencies": {
+    "eslint": ">=6.5.1 <7"
+  }
+}
