+++ src/eslint-runner/src/shouldLintAll.js
