+++ src/eslint-runner/src/stylish.js
