+++ src/eslint-runner/src/verifyPackageTree.js
